refactor(storage): move storage-related code to new package
- Moved `storage` related files from `packages/cosec/src/storage` to `packages/storage/src`.- Updated import paths in `packages/cosec` to use the new `@ahoo-wang/fetcher-storage` package.
- Removed redundant exports and imports in `packages/cosec/src/index.ts` and other affected files.
- Added `@ahoo-wang/fetcher-storage` as a dependency in `packages/cosec/package.json`.This refactoring improves modularity and separation of concerns by isolating storage-related functionality into its own package.
