+++ packages/storage/package.json
@@ -1,25 +1,28 @@
 {
-  "name": "@ahoo-wang/fetcher-cosec",
+  "name": "@ahoo-wang/fetcher-storage",
   "version": "1.6.6",
-  "description": "CoSec authentication integration for Fetcher HTTP client with enterprise-grade security features. Provides automatic token management, device ID persistence, and request tracking.",
+  "description": "Ultra-lightweight HTTP client with built-in path parameters, query parameters, and Axios-like API. The foundation of the Fetcher ecosystem with powerful interceptor system and LLM streaming API support.",
   "keywords": [
     "fetch",
     "http",
     "client",
-    "cosec",
-    "authentication",
-    "security",
-    "enterprise",
-    "token-management",
-    "device-id"
+    "axios",
+    "ultra-lightweight",
+    "modular",
+    "typescript",
+    "interceptor",
+    "llm",
+    "streaming",
+    "sse",
+    "server-sent-events"
   ],
   "author": "Ahoo-Wang",
   "license": "Apache-2.0",
-  "homepage": "https://github.com/Ahoo-Wang/fetcher/tree/master/packages/cosec",
+  "homepage": "https://github.com/Ahoo-Wang/fetcher/tree/master/packages/fetcher",
   "repository": {
     "type": "git",
     "url": "https://github.com/Ahoo-Wang/fetcher.git",
-    "directory": "packages/cosec"
+    "directory": "packages/storage"
   },
   "bugs": {
     "url": "https://github.com/Ahoo-Wang/fetcher/issues"
@@ -44,11 +47,6 @@
     "lint": "eslint . --fix",
     "clean": "rm -rf dist",
     "analyze": "npx vite-bundle-analyzer -p auto"
-  },
-  "dependencies": {
-    "@ahoo-wang/fetcher": "workspace:*",
-    "@ahoo-wang/fetcher-storage": "workspace:*",
-    "nanoid": "^5.1.5"
   },
   "devDependencies": {
     "@vitest/coverage-v8": "^3.2.4",
